Implementing test framework
Started creating an x86 instruction generator, capable of generating *all* x86
instructions.

Found and patched a few bugs in lend as a result of initial testing.

At the moment all instructions with 1 byte opcodes are generated, and around
half of all instructions with 2 byte opcodes are being generated. This is a
worki in progress.